Removal of the old fashioned schema classes
It also removes the tests
