+++ deprecated/frontend/schema/class-schema-article.php
@@ -1,0 +1,70 @@
+<?php
+/**
+ * WPSEO plugin file.
+ *
+ * @package WPSEO\Frontend\Schema
+ */
+
+/**
+ * Returns schema Article data.
+ *
+ * @deprecated xx.x
+ *
+ * @since 10.2
+ */
+class WPSEO_Schema_Article implements WPSEO_Graph_Piece {
+
+	/**
+	 * WPSEO_Schema_Article constructor.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 */
+	public function __construct() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+	}
+
+	/**
+	 * Determines whether or not a piece should be added to the graph.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @return bool
+	 */
+	public function is_needed() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return false;
+	}
+
+	/**
+	 * Returns Article data.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @return array $data Article data.
+	 */
+	public function generate() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return array();
+	}
+
+	/**
+	 * Determines whether a given post type should have Article schema.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @param string $post_type Post type to check.
+	 *
+	 * @return bool True if it has article schema, false if not.
+	 */
+	public static function is_article_post_type( $post_type = null ) {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return false;
+	}
+}
