+++ deprecated/frontend/schema/class-schema-breadcrumb.php
@@ -1,0 +1,50 @@
+<?php
+/**
+ * WPSEO plugin file.
+ *
+ * @package WPSEO\Frontend\Schema
+ */
+
+/**
+ * Returns schema Breadcrumb data.
+ *
+ * @deprecated xx.x
+ *
+ * @since 10.2
+ */
+class WPSEO_Schema_Breadcrumb implements WPSEO_Graph_Piece {
+
+	/**
+	 * WPSEO_Schema_Breadcrumb constructor.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 */
+	public function __construct() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+	}
+
+	/**
+	 * Determine if we should add a breadcrumb attribute.
+	 *
+	 * @return bool
+	 */
+	public function is_needed() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return false;
+	}
+
+	/**
+	 * Returns Schema breadcrumb data to allow recognition of page's position in the site hierarchy.
+	 *
+	 * @link https://developers.google.com/search/docs/data-types/breadcrumb
+	 *
+	 * @return bool|array Array on success, false on failure.
+	 */
+	public function generate() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return array();
+	}
+}
