+++ deprecated/frontend/schema/class-schema-faq-questions.php
@@ -1,0 +1,44 @@
+<?php
+/**
+ * WPSEO plugin file.
+ *
+ * @package WPSEO\Frontend\Schema
+ */
+
+/**
+ * Returns a question object for each question in an FAQ block.
+ *
+ * @deprecated xx.x
+ *
+ * @since 11.1
+ */
+class WPSEO_Schema_FAQ_Questions {
+
+	/**
+	 * WPSEO_Schema_FAQ_Questions constructor.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @param array                 $data    Our schema graph.
+	 * @param WP_Block_Parser_Block $block   The FAQ block of this type.
+	 * @param WPSEO_Schema_Context  $context A value object with context variables.
+	 */
+	public function __construct( $data, $block, $context ) {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+	}
+
+	/**
+	 * Find an image based on its URL and generate a Schema object for it.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @return array The Schema with Questions added.
+	 */
+	public function generate() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return array();
+	}
+}
