+++ deprecated/frontend/schema/class-schema-faq.php
@@ -8,84 +8,69 @@
 /**
  * Returns schema FAQ data.
  *
+ * @deprecated xx.x
+ *
  * @since 11.3
  */
 class WPSEO_Schema_FAQ implements WPSEO_Graph_Piece {
 
 	/**
-	 * Determine whether this graph piece is needed or not.
-	 *
-	 * @var bool
-	 */
-	private $is_needed = false;
-
-	/**
-	 * The FAQ blocks on the current page.
-	 *
-	 * @var array
-	 */
-	private $blocks;
-
-	/**
-	 * A value object with context variables.
-	 *
-	 * @var WPSEO_Schema_Context
-	 */
-	private $context;
-
-	/**
 	 * WPSEO_Schema_FAQ constructor.
 	 *
-	 * @param WPSEO_Schema_Context $context A value object with context variables.
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
 	 */
-	public function __construct( WPSEO_Schema_Context $context ) {
-		$this->context = $context;
-
-		add_action( 'wpseo_pre_schema_block_type_yoast/faq-block', array( $this, 'prepare_schema' ), 10, 1 );
-		add_filter( 'wpseo_schema_block_yoast/faq-block', array( $this, 'render_schema_questions' ), 10, 3 );
+	public function __construct() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
 	}
 
 	/**
 	 * If this fires, we know there's an FAQ block ont he page, so filter the page type.
 	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
 	 * @param array $blocks The blocks of this type on the current page.
 	 */
 	public function prepare_schema( $blocks ) {
-		$this->blocks    = $blocks;
-		$this->is_needed = true;
-		add_filter( 'wpseo_schema_webpage_type', array( $this, 'change_schema_page_type' ) );
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
 	}
 
 	/**
 	 * Change the page type to an array if it isn't one, include FAQPage.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
 	 *
 	 * @param array|string $page_type The page type.
 	 *
 	 * @return array $page_type The page type that's now an array.
 	 */
 	public function change_schema_page_type( $page_type ) {
-		if ( ! is_array( $page_type ) ) {
-			$page_type = array( $page_type );
-		}
-		$page_type[] = 'FAQPage';
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
 
-		return $page_type;
+		return array();
 	}
 
 	/**
 	 * Render a list of questions, referencing them by ID.
 	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
 	 * @return array $data Our Schema graph.
 	 */
 	public function generate() {
-		$question_list = new WPSEO_Schema_FAQ_Question_List( $this->blocks, $this->context );
-		$graph         = $question_list->generate();
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
 
-		return $graph;
+		return array();
 	}
 
 	/**
 	 * Add the Questions in our FAQ blocks as separate pieces to the graph.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
 	 *
 	 * @param array                 $graph   Schema data for the current page.
 	 * @param WP_Block_Parser_Block $block   The block data array.
@@ -94,18 +79,22 @@ class WPSEO_Schema_FAQ implements WPSEO_Graph_Piece {
 	 * @return array $data Our Schema graph.
 	 */
 	public function render_schema_questions( $graph, $block, $context ) {
-		$questions = new WPSEO_Schema_FAQ_Questions( $graph, $block, $context );
-		$graph     = $questions->generate();
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
 
-		return $graph;
+		return array();
 	}
 
 	/**
 	 * Determines whether or not a piece should be added to the graph.
 	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
 	 * @return bool
 	 */
 	public function is_needed() {
-		return $this->is_needed;
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return false;
 	}
 }
