+++ deprecated/frontend/schema/class-schema-howto.php
@@ -1,0 +1,71 @@
+<?php
+/**
+ * WPSEO plugin file.
+ *
+ * @package WPSEO\Frontend\Schema
+ */
+
+/**
+ * Returns schema FAQ data.
+ *
+ * @deprecated xx.x
+ *
+ * @since 11.5
+ */
+class WPSEO_Schema_HowTo implements WPSEO_Graph_Piece {
+
+	/**
+	 * WPSEO_Schema_FAQ constructor.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 */
+	public function __construct() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+	}
+
+	/**
+	 * Renders a list of questions, referencing them by ID.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @return array $data Our Schema graph.
+	 */
+	public function generate() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return array();
+	}
+
+	/**
+	 * Renders the How-To block into our graph.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @param array $graph Our Schema data.
+	 * @param array $block The How-To block content.
+	 *
+	 * @return mixed
+	 */
+	public function render( $graph, $block ) {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return array();
+	}
+
+	/**
+	 * Determines whether or not a piece should be added to the graph.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @return bool
+	 */
+	public function is_needed() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return false;
+	}
+}
