+++ deprecated/frontend/schema/class-schema-organization.php
@@ -1,0 +1,54 @@
+<?php
+/**
+ * WPSEO plugin file.
+ *
+ * @package WPSEO\Frontend\Schema
+ */
+
+/**
+ * Returns schema Organization data.
+ *
+ * @deprecated xx.x
+ *
+ * @since 10.2
+ */
+class WPSEO_Schema_Organization implements WPSEO_Graph_Piece {
+
+	/**
+	 * WPSEO_Schema_Organization constructor.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 */
+	public function __construct() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+	}
+
+	/**
+	 * Determines whether an Organization graph piece should be added.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @return bool
+	 */
+	public function is_needed() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return false;
+	}
+
+	/**
+	 * Returns the Organization Schema data.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @return array $data The Organization schema.
+	 */
+	public function generate() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return array();
+	}
+}
