+++ deprecated/frontend/schema/class-schema-person.php
@@ -1,0 +1,54 @@
+<?php
+/**
+ * WPSEO plugin file.
+ *
+ * @package WPSEO\Frontend\Schema
+ */
+
+/**
+ * Returns schema Person data.
+ *
+ * @deprecated xx.x
+ *
+ * @since 10.2
+ */
+class WPSEO_Schema_Person implements WPSEO_Graph_Piece {
+
+	/**
+	 * WPSEO_Schema_Person constructor.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 */
+	public function __construct() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+	}
+
+	/**
+	 * Determine whether we should return Person schema.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @return bool
+	 */
+	public function is_needed() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return false;
+	}
+
+	/**
+	 * Returns Person Schema data.
+	 *
+	 * @codeCoverageIgnore
+	 * @deprecated xx.x
+	 *
+	 * @return bool|array Person data on success, false on failure.
+	 */
+	public function generate() {
+		_deprecated_function( __METHOD__, 'WPSEO xx.x' );
+
+		return array();
+	}
+}
