+++ deprecated/frontend/schema/interface-wpseo-graph-piece.php
@@ -9,6 +9,8 @@ if ( ! interface_exists( 'WPSEO_Graph_Piece' ) ) {
 	/**
 	 * An interface for registering Schema Graph Pieces.
 	 *
+	 * @deprecated xx.x
+	 *
 	 * @since 10.2
 	 */
 	interface WPSEO_Graph_Piece {
